requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.3
mysql-connector-python==8.1.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
bcrypt==4.0.1
redis==5.0.1
requests==2.31.0

